% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kafka_operations.R
\name{delete_cluster}
\alias{delete_cluster}
\title{Deletes the MSK cluster specified by the Amazon Resource Name (ARN) in the request}
\usage{
delete_cluster(
  ClusterArn = "string",
  CurrentVersion = "string"
)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{CurrentVersion}{The current version of the MSK cluster.}
}
\description{
Deletes the MSK cluster specified by the Amazon Resource Name (ARN) in the request.
}
\examples{

}
