% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kafka_operations.R
\name{describe_cluster}
\alias{describe_cluster}
\title{Returns a description of the MSK cluster whose Amazon Resource Name (ARN) is specified in the request}
\usage{
describe_cluster(ClusterArn)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}
}
\description{
Returns a description of the MSK cluster whose Amazon Resource Name (ARN) is specified in the request.
}
\section{Accepted Parameters}{
\preformatted{describe_cluster(
  ClusterArn = "string"
)
}
}

