% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kafka_operations.R
\name{get_bootstrap_brokers}
\alias{get_bootstrap_brokers}
\title{A list of brokers that a client application can use to bootstrap}
\usage{
get_bootstrap_brokers(ClusterArn)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}
}
\description{
A list of brokers that a client application can use to bootstrap.
}
\section{Accepted Parameters}{
\preformatted{get_bootstrap_brokers(
  ClusterArn = "string"
)
}
}

