% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kafka_operations.R
\name{list_nodes}
\alias{list_nodes}
\title{Returns a list of the broker nodes in the cluster}
\usage{
list_nodes(ClusterArn, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{ClusterArn}{[required] The Amazon Resource Name (ARN) that uniquely identifies the cluster.}

\item{MaxResults}{The maximum number of clusters to return in the response. If there are more clusters, the response includes a NextToken parameter.}

\item{NextToken}{The paginated results marker. When the result of a ListClusters operation is truncated, the call returns NextToken in the response. To get another batch of clusters, provide this token in your next request.}
}
\description{
Returns a list of the broker nodes in the cluster.
}
\section{Accepted Parameters}{
\preformatted{list_nodes(
  ClusterArn = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

