% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kafka_package.R
\docType{package}
\name{paws.kafka-package}
\alias{paws.kafka}
\alias{paws.kafka-package}
\title{paws.kafka: Managed Streaming for Kafka}
\description{
The operations for managing an Amazon MSK cluster.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
